+++ includes/ofdm/phasereference.h
@@ -40,22 +40,22 @@ public:
 	                                         uint8_t,
 	                                         int16_t);
 			~phaseReference		(void);
-	int32_t		findIndex		(std::vector<std::complex<float>>, int);
-	int16_t		estimate_CarrierOffset	(std::vector<std::complex<float>>);
-	float		estimate_FrequencyOffset (std::vector<std::complex<float>>);
+	int32_t		findIndex		(std::vector<std::complex<double>>, int);
+	int16_t		estimate_CarrierOffset	(std::vector<std::complex<double>>);
+	double		estimate_FrequencyOffset (std::vector<std::complex<double>>);
 //
 //	This one is used in the ofdm decoder
-	std::vector<std::complex<float>> refTable;
+	std::vector<std::complex<double>> refTable;
 private:
 	dabParams	params;
 	fftHandler	my_fftHandler;
-	std::vector<float> phaseDifferences;
+	std::vector<double> phaseDifferences;
 	int16_t		threshold;
 	int16_t		diff_length;
 	int32_t		T_u;
 	int16_t		carriers;
 
-	std::complex<float>	*fft_buffer;
+	std::complex<double>	*fft_buffer;
 	int32_t		fft_counter;
 	int32_t		framesperSecond;	
 	int32_t		displayCounter;
