+++ includes/scopes-qwt6/iqdisplay.h
@@ -51,12 +51,12 @@ Q_OBJECT
 public:
 	IQDisplay		(QwtPlot *, int16_t);
 	~IQDisplay		(void);
-void	DisplayIQ		(std::complex<float> *, float);
+void	DisplayIQ		(std::complex<double> *, double);
 private:
 	int32_t			x_amount;
 	std::vector<double>	plotData;
 	std::vector<double>	plot2;
-	std::vector<complex<float> > Points;
+	std::vector<complex<double> > Points;
 	QwtPlot			*plotgrid;
 	int			_OutputRate;
 	int			Radius;
